# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "API for aggregating news from multiple sources",
        "contact": {
            "email": "developer@innoscripta.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/v1/articles/search": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Search articles",
                "description": "Search articles with filters",
                "operationId": "c0ad0695ef4aae27e540e5faa1fc87f2",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "bitcoin"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "technology"
                        }
                    },
                    {
                        "name": "source_id",
                        "in": "query",
                        "description": "Source ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "description": "Author name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "John Doe"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Start date (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-01-01"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "End date (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-12-31"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Results per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Article"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get single article",
                "description": "Get article by ID",
                "operationId": "bf9724a6309b55de3ed51502e1a15814",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Article": {
                "title": "Article Model",
                "description": "Detailed Article object",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the article",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "description": "The title of the article",
                        "type": "string"
                    },
                    "content": {
                        "description": "The main body content of the article",
                        "type": "string"
                    },
                    "published_at": {
                        "description": "The date and time the article was published",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "title": "A Great Blog Post",
                    "content": "This is the content of the article.",
                    "published_at": "2024-11-08T10:00:00.000000Z"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}